Fix feature environments not being given urls.
Environment urls are not double expanded, meaning if we use a single
variable $FEATURE_DOMAIN which itself includes $CI_ENVIRONMENT_SLUG,
the url will not be complete and gitlab will ignore it. This results
in noone being able to discover the url for a feature branch except by
logging in to the development server themselves.

Beginning with this change we expect the projects that use this to set
an explicit value for FEATURE_DOMAIN, and we will prepend the unique
environment slug piece here in the central ci configs.